Social | Migrate shares data from initial state to new script data
