+++ projects/js-packages/publicize-components/src/social-store/reducer/index.js
@@ -1,12 +1,10 @@
 import { combineReducers } from '@wordpress/data';
 import connectionData from './connection-data';
 import jetpackSettings from './jetpack-settings';
-import sharesData from './shares-data';
 import siteData from './site-data';
 import socialImageGeneratorSettings from './social-image-generator-settings';
 
 const reducer = combineReducers( {
-	sharesData,
 	siteData,
 	connectionData,
 	jetpackSettings,
