--- projects/js-packages/publicize-components/src/social-store/reducer/shares-data.js
@@ -1,5 +1,0 @@
-const sharesData = ( state = {} ) => {
-	return state;
-};
-
-export default sharesData;
