+++ projects/js-packages/publicize-components/src/social-store/selectors/index.js
@@ -1,6 +1,5 @@
 import * as connectionDataSelectors from './connection-data';
 import jetpackSettingSelectors from './jetpack-settings';
-import * as sharesData from './shares-data';
 import siteDataSelectors from './site-data';
 import socialImageGeneratorSettingsSelectors from './social-image-generator-settings';
 
@@ -8,7 +7,6 @@ const selectors = {
 	...siteDataSelectors,
 	...connectionDataSelectors,
 	...jetpackSettingSelectors,
-	...sharesData,
 	...socialImageGeneratorSettingsSelectors,
 	userConnectionUrl: state => state.userConnectionUrl,
 	useAdminUiV1: state => state.useAdminUiV1,
