+++ projects/packages/publicize/src/class-publicize-script-data.php
@@ -12,6 +12,7 @@ use Automattic\Jetpack\Connection\Manager;
 use Automattic\Jetpack\Current_Plan;
 use Automattic\Jetpack\Publicize\Publicize_Utils as Utils;
 use Automattic\Jetpack\Status\Host;
+use Jetpack_Options;
 
 /**
  * Publicize_Script_Data class.
@@ -84,6 +85,7 @@ class Publicize_Script_Data {
 			'is_publicize_enabled' => Utils::is_publicize_active(),
 			'feature_flags'        => self::get_feature_flags(),
 			'supported_services'   => array(),
+			'shares_data'          => array(),
 		);
 
 		if ( ! Utils::is_publicize_active() ) {
@@ -102,10 +104,10 @@ class Publicize_Script_Data {
 			array(
 				'api_paths'          => self::get_api_paths(),
 				'supported_services' => self::get_supported_services(),
+				'shares_data'        => self::get_shares_data(),
 				/**
 				 * 'store'       => self::get_store_script_data(),
 				 * 'urls'        => self::get_urls(),
-				 * 'shares_data' => self::get_shares_data(),
 				 */
 			)
 		);
@@ -151,6 +153,15 @@ class Publicize_Script_Data {
 		}
 
 		return Current_Plan::supports( 'social-' . $feature );
+	}
+
+	/**
+	 * Get the shares data.
+	 *
+	 * @return ?array
+	 */
+	public static function get_shares_data() {
+		return self::publicize()->get_publicize_shares_info( Jetpack_Options::get_option( 'id' ) );
 	}
 
 	/**
